src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asyncpg==0.31.0
attrs==25.4.0
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
typing_extensions==4.15.0
yarl==1.22.0
